refactored some sqlalchemy code to reuse in cassandra backend
