+++ frontera/contrib/backends/cassandra/models.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-import uuid
+from cassandra.cqlengine.columns import (BigInt, Boolean, Counter, DateTime,
+                                         Float, Integer, Map, SmallInt, Text,
+                                         UserDefinedType)
 from cassandra.cqlengine.models import Model
 from cassandra.cqlengine.usertype import UserType
-from cassandra.cqlengine.columns import Map, Text, Float, Integer, DateTime, UserDefinedType, Counter, Boolean, \
-    SmallInt, BigInt
 
 
 class Meta(UserType):
